Pixel Paintballers v1.3.1-alpha
Apologies for all this, uploading to GitHub during late hours isn't a
very good practice. That aside,
Changes include:
- Statistics system
- Saving of player's customisation and statistics

Following v1.3.1-alpha, the development of Pixel Paintballers  will now
progress to Beta, where the main focus will be multiplayer
functionality. When these releases are uploaded, it can be noticed that
the game's look has been reverted. This is to test the functionality of
the system. After that, the game will regain its current look with the
system fully implemented.