Pixel Paintballers v1.3.2-alpha
Although the main focus is on beta development, I felt the need to clean up the statistic management, mainly because I had only setup the bare bones of it in v1.3.1-alpha.
Changes include:
- Game data is now all saved in one file, "Data.txt". (The system will not read the old text files, so customisations will be lost and stats reset.)
- Time played statistic has been revamped, making it look more appealing.
- Total kills now accumulate after a completed game.
- Homing paintball has been slowed down in order to make it less daunting.
- New sound plays when trying to access the multiplayer menu, as it is not in alpha releases.

Development is now heavily focused on Beta, but there's still quite a lot of bugs to clean out before I can even give the first release.