Pixel Paintballers v1.0-beta
Here it is. At long last we have the first beta release for Pixel Paintballers, focusing on multiplayer functionality. This first release consists of a 4-player free-for-all gamemode. For the time being, there are only 2 major updates planned for beta, with this being the first. The second update will consist of a team survival gamemode.