Pixel Paintballers v1.1-beta
Changes include:
 - Multiplayer PVE gamemode, "Team survival"
 - Added multiplayer menu that will be brought over to the full game.

The next update v1.1.1-beta will focus on adding spectators and possibly fixing a known bug that causes some de-synchronisation between the clients and the server in terms of enemy location and management. This will usually occur in the form of no spawning of enemies during a multiplayer game. The best way to deal with this issue currently is to shoot paintballs across the whole screen to eliminate any enemies missed.